refactor: Migrate write-project-content to typescript (LPS-119618)
